requirements.txt
absl-py==0.8.1
alabaster==0.7.12
astor==0.8.0
auto-sklearn==0.6.0
Babel==2.7.0
certifi==2019.9.11
chardet==3.0.4
ConfigSpace==0.4.11
cycler==0.10.0
Cython==0.29.13
deap==1.3.0
docutils==0.15.2
gast==0.2.2
google-pasta==0.1.7
grpcio==1.24.1
h5py==2.10.0
idna==2.8
imagesize==1.1.0
Jinja2==2.10.3
joblib==0.14.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
liac-arff==2.4.0
lockfile==0.12.2
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
nose==1.3.7
numpy==1.17.2
opt-einsum==3.1.0
packaging==19.2
pandas==0.25.1
progressbar2==3.47.0
protobuf==3.10.0
psutil==5.6.3
pydot==1.4.1
Pygments==2.4.2
pynisher==0.5.0
pyparsing==2.4.2
pyrfr==0.8.0
python-dateutil==2.8.0
python-utils==2.3.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
smac==0.8.0
snowballstemmer==2.0.0
Sphinx==2.2.0
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
stopit==1.1.2
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.0
termcolor==1.1.0
TPOT==0.10.2
tqdm==4.36.1
typing==3.7.4.1
update-checker==0.16
urllib3==1.25.6
Werkzeug==0.16.0
wrapt==1.11.2
xgboost==0.90
